Correct typos; abbreviations. Calculate by onlyNeededVector.
Eliminate most abnormal negative scores. But still some unsolved.
(not typos, in newVector.txt)
this is an abnormal cos:  1.00607456992 ('weird', 'position')
this is an abnormal cos:  1.00607456992 ('weird', 'position')
this is an abnormal cos:  1.02688875348 ('position', 'white dot centered')
this is an abnormal cos:  1.02688875348 ('white dot centered', 'position')
this is an abnormal cos:  1.01013007563 ('incomplete', 'concave feature')